RT 140734 - Adding large durations. First commit
This commit fixes the reported problem by removing code in
add_duration() that adjusts the date of the object if the initial date
and final date were not both Julian or both Gregorian.

Unfortunately this causes test failures in t/08math.t, and addition of
large intervals no longer round-trips. I am committing anyway to have a
baseline to work from.