RT 140734 - Adding large durations. Second commit
Fixed test failures in t/08math.t introduced by first commit.

This was done by processing the components of DateTime::Duration
separately, in the order documented by DateTime: days, months, minutes,
seconds, nanoseconds. The adjustment for Gregorian deviation that was
removed in the original code was applied only to the months.

It seems to me (IM(NS)HO) that the adjustment does not make a huge
amount of sense -- or at least that there are things to do that are at
least as sane as this.